Added code to perform a multi-threaded Tiff export
Created a new wrapper class for the TiffWriter that operates in a separate thread. The plane bytes are passed using a thread-safe BlockingQueue.

Added code to the ExporterI class to use the background thread worker if the file to be written is above a certain size. Added timings for the methods that open the image plane from OMERO, save the plane using BioFormats and the time that the master thread has to wait for the background writer to finish.

Overall timings for exporting different images are:

MB     Single      Multi       Ratio
0.5    46.03       44.42       0.97
1      79.78       79.11       0.99
2      112.59      106.21      0.94
4      217.53      210.07      0.97
8      436.03      421.24      0.97
16     864.59      840.96      0.97
32     1721.95     1675.20     0.97
64     3441.70     3361.51     0.98
128    6879.39     6727.62     0.98
256    13650.14    13358.87    0.98

The background writing therefore adds a 2-3% increase in performance.

Given the low level of performance increase I suggested that the patch is not worth pursueing further:
http://www.openmicroscopy.org/community/viewtopic.php?f=6&t=814&sid=a870884013f0657bd3a079d1ca16e41d#p3107